The construction has been completed,PR to Dev111. (#17)
* 重构main模块

* 重构main模块

* 修正main模块

* 修正main模块

* 重构furrymodule，更改菜单功能可用性

* 重构文件校验函数并同步

* 修复错误的默认传入数据

* 修复furrybar bug

* 重构furrybar主函数。

* 更新 .env.example 文件

* fix: 修复不响应`签到`命令的bug

* 重构bug反馈功能、结婚系统、菜单系统

---------
