+++ .env.example
@@ -1,7 +1,6 @@
 # =============================================================
 #  基础配置
 # =============================================================
-
 HOST=127.0.0.1
 PORT=9090
 DRIVER=~fastapi+~httpx+~websockets
@@ -19,8 +18,10 @@ SUPERUSERS=["QQ号1","QQ号2"]# 最高权限用户
 SENSITIVEADMIN=["QQ号1","QQ号2"]# 敏感词管理员
 
 # =============================================================
-#  插件配置
+#  官方插件配置
 # =============================================================
+#  nonebot_plugin_mc_server_status 的配置
+MC_STATUS_ADMIN_QQNUM=[QQ号1, QQ号2]
 
 # nonebot_plugin_localstore 的配置
 LOCALSTORE_USE_CWD=True
@@ -32,19 +33,20 @@ NCM_ANONYMOUS=True
 NCM_CARD_SIGN_URL=https://llob.linyuchen.net/sign/music
 # NCM_ILLEGAL_CMD_FINISH=True
 
+# =============================================================
+#  自定义插件配置
+# =============================================================
 #  Furry 插件配置
 FURRY_TOKEN="your_furry_token"
 FURRY_USER="your_furry_username"
 FURRY_PASSWORD="your_furry_password"
 FURRY_AIKEY="Bearer your_ai_api_key"
 FURRY_SPECIAL_EVENTS=["hifurry","极兽聚","兽行迹"]
+FURRY_AI_BASEURL="your_baseurl"
 
 #  狼人杀配置
 MIN_PLAYERS_LRS=2
 MAX_PLAYERS_LRS=12
-
-#  nonebot_plugin_mc_server_status的配置
-MC_STATUS_ADMIN_QQNUM=[QQ号1, QQ号2]
 
 #  GitHub 集成配置
 GITHUB_TOKEN="your_github_personal_access_token"
