+++ markdown/furry_system.md
@@ -111,5 +111,5 @@
 | 可用 | 功能                                                                          | 语法示例/用法                                                     | 效果                                                                         |
 |----|-----------------------------------------------------------------------------|-------------------------------------------------------------|----------------------------------------------------------------------------|
 | ✅  | 修改图片#<数字 id>#<名字/留言/类型/如果要修改图片请留空>#<要修改的内容或图片>                              | `修改图片#9162#名字#海龟`<br>`修改图片#9162##<image图片>`                 | 修改<em>通过凌辉Bot</em>上传的图片信息                                                  |
-| ⚠️ | 审核列表/待审核列表/上传列表                                                             | `直接发送`                                                      | 输出通过`投图命令`的<em>未审核</em>的图片列表                                               |
-| ⚠️ | 同意上传#<数字 条数><br>同意上载#<数字 条数><br><em>或者</em><br>拒绝上传#<数字 条数><br>拒绝上载#<数值 条数> | `同意上传#1`<br>`同意上载#1`<br><em>或者</em><br>`拒绝上传#1`<br>`拒绝上载#1` | 操作<em>**是否批准**</em>`第1张`图片上传请求<br><em>如果批准，该请求将上传至`兽云祭API`并由其管理员进行复核。</em> |
+| ✅ | 审核列表/待审核列表/上传列表                                                             | `直接发送`                                                      | 输出通过`投图命令`的<em>未审核</em>的图片列表                                               |
+| ✅ | 同意上传#<数字 条数><br>同意上载#<数字 条数><br><em>或者</em><br>拒绝上传#<数字 条数><br>拒绝上载#<数值 条数> | `同意上传#1`<br>`同意上载#1`<br><em>或者</em><br>`拒绝上传#1`<br>`拒绝上载#1` | 操作<em>**是否批准**</em>`第1张`图片上传请求<br><em>如果批准，该请求将上传至`兽云祭API`并由其管理员进行复核。</em> |
