+++ src/__init__.py
@@ -1,5 +1,5 @@
 from nonebot import require
 
 
-# require("nonebot_plugin_orm")
-# require("nonebot_plugin_alconna")
+require("nonebot_plugin_orm")
+require("nonebot_plugin_alconna")
