+++ src/plugins/furrymodule/furry/__init__.py
@@ -1,0 +1,2 @@
+from .furry import *
+from .upload import *
