+++ src/plugins/furrymodule/furry/upload/__init__.py
@@ -1,0 +1,1 @@
+from .upload import *
