+++ src/plugins/furrymodule/furrybar/__init__.py
@@ -1,0 +1,1 @@
+from .furrybar import *
