+++ src/plugins/marry_system/__init__.py
@@ -1,0 +1,1 @@
+from .marry import *
