+++ src/plugins/menu/__init__.py
@@ -1,0 +1,1 @@
+from .menu import *
