+++ src/plugins/report_bug/__init__.py
@@ -1,0 +1,1 @@
+from .report_bug import *
